get_wgfbit_data2 {icesVMS}	R Documentation

Download swept area ratio, landings and value map data

Description

Download a data.frame of surface and subsurface swept area ratio by c-square for a given ICES ecoregion, year and gear code.

Usage

get_wgfbit_data2(ecoregion, year)

Arguments

ecoregion
ICES ecoregion

year
which year to select

Value

a data.frame with a WKT column for the c-square polygons

[Package icesVMS version 1.0.0.9006 Index]