icesVMS-package {icesVMS}	R Documentation

Functions to link to the ICES VMS and logbook database webservices.

Description

Functions to link to the ICES VMS and logbook database webservices to allow users to download summaries and data products from the ICES VMS and logbook database.

Details

Download data:

get_benthis_parameters 	 download gear contact model parameters

get_metier_lookup 	 download metier to gear category lookup table

get_vms 	 download VMS data

Authentication:

update_token 	 update the token for the current project

Author(s)

Colin Millar.

References

ICES web services: https://www.ices.dk/marine-data/tools/Pages/WebServices.aspx.

[Package icesVMS version 1.0.0.9006 Index]