update_token {icesVMS}	R Documentation

Update user token for project

Description

Must be run in interactive mode. User is promted for a password, and if successful a Javascript Web Token (JWT) is saved to the project root (current working directory). The token will last for 1 hour, after which the user needs to update the token using this function again.

Usage

update_token(user)

Arguments

user
a character string giving the users ICES username

Value

TRUE or FALSE depenting on the success of the operation

[Package icesVMS version 1.0.0.9006 Index]